idswidget/requirements.txt
python-fasthtml
requests
pydantic
